# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CornerLab API
  description: >
    Plataforma de inteligência estatística, backtesting e gestão de estratégia
    para o mercado de escanteios. A API nunca recomenda apostas — apenas
    organiza dados históricos, calcula estatísticas e simula cenários.
    O Módulo de Inteligência Estatística (grupo `intelligence`) segue as mesmas
    restrições: nenhum endpoint prevê resultados futuros, promete lucro ou
    recomenda mercados — todo indicador é derivado de dados históricos
    armazenados, com rastreabilidade (liga/temporada/jogos analisados/data de
    atualização) em `meta` nas respostas.
  version: "0.2.0"
servers:
  - url: /api/v1
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Cadastrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { description: Usuário criado }
  /auth/login:
    post:
      tags: [auth]
      summary: Autenticar usuário
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: Login efetuado }
...
  /leagues/{id}/seasons:
    get:
      tags: [catalog]
      summary: Listar temporadas de um campeonato
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Lista de temporadas }
  /teams:
    get:
      tags: [catalog]
      summary: Listar ou pesquisar equipes
      parameters:
        - in: query
          name: league_id
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200': { description: Lista de equipes }
...
  /bets:
    get:
      tags: [bets]
      summary: Listar histórico de apostas (autenticado)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de apostas }
    post:
      tags: [bets]
      summary: Registrar aposta (autenticado, Módulo 5)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BetRequest' }
      responses:
        '201': { description: Aposta registrada }
  /bets/dashboard:
    get:
      tags: [bets]
      summary: Dashboard financeiro (autenticado, Módulo 5/8)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Métricas financeiras agregadas }
  /bets/{id}:
    delete:
      tags: [bets]
      summary: Excluir aposta (autenticado)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Aposta excluída }
...
  /alerts:
    get:
      tags: [alerts]
      summary: Listar regras de alerta do usuário autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lista de regras de alerta }
    post:
      tags: [alerts]
      summary: Criar regra de alerta inteligente (Módulo 7)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAlertRequest' }
      responses:
        '201': { description: Regra criada }
  /alerts/{id}/evaluate:
    post:
      tags: [alerts]
      summary: Avaliar uma regra de alerta agora
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: "Resultado da avaliação (disparada ou não, com jogos/equipes que casaram)" }
  /alerts/{id}:
    delete:
      tags: [alerts]
      summary: Excluir regra de alerta
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '204': { description: Regra excluída }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string, minLength: 6 }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
...
    BetRequest:
      type: object
      required: [match_label, market, odd, stake, event_date]
      properties:
        match_label: { type: string }
        league_id: { type: integer, nullable: true }
        market: { type: string, example: "Mais de 9.5 escanteios" }
        odd: { type: number }
        stake: { type: number }
        status: { type: string, enum: [pending, won, lost, void] }
        event_date: { type: string, example: "2025-05-10" }
...
    AlertDefinition:
      type: object
      required: [type, league_id]
      description: >
        `type=team_frequency` avalia a frequência de uma equipe acima de um
        threshold de escanteios nos últimos N jogos. `type=opponent_average`
        avalia se a média sofrida por um adversário está acima de um mínimo.
      properties:
        type: { type: string, enum: [team_frequency, opponent_average] }
        league_id: { type: integer }
        team_id: { type: integer, nullable: true }
        last_n_games: { type: integer, example: 10 }
        corners_threshold: { type: integer, example: 9 }
        threshold_pct: { type: number, example: 70, description: "% mínimo de jogos acima do threshold (team_frequency)" }
        min_avg_conceded: { type: number, example: 6.5, description: "Média mínima sofrida (opponent_average)" }
    CreateAlertRequest:
      type: object
      required: [name, definition]
      properties:
        name: { type: string, example: "Flamengo acima de 9.5 em 70% dos jogos" }
        definition: { $ref: '#/components/schemas/AlertDefinition' }